requirements.txt
loguru~=0.7.2
pycryptodome~=3.20.0
mysql~=0.0.3